Updates the simann library to 2003 C++
This commit also changes the build system to CMake and add README and LICENSE
files to the project as suggested by GitHub.

Most of the files in the original project were updated to use modern headers
and left.  However, the erstream.{hpp,cxx} was too complex and depended on
several nonstandard C++ features that are no longer available.  It has been
replaced by a simpler class with similar purpose in the files errorstream.*.